# src/presentation/api/docs/swagger.yaml
basePath: /v1
definitions:
...
  dto.AddContainerProfile:
    properties:
      baseSpecs:
        $ref: "#/definitions/valueObject.ContainerSpecs"
      hostMinCapacityPercent:
        type: number
      maxSpecs:
        $ref: "#/definitions/valueObject.ContainerSpecs"
      name:
        type: string
      scalingIntervalSecs:
        type: integer
      scalingMaxDurationSecs:
        type: integer
      scalingPolicy:
        $ref: "#/definitions/valueObject.ScalingPolicy"
      scalingThreshold:
        type: integer
    type: object
...
  dto.UpdateContainerProfile:
    properties:
      baseSpecs:
        $ref: "#/definitions/valueObject.ContainerSpecs"
      hostMinCapacityPercent:
        type: number
      id:
        type: integer
      maxSpecs:
        $ref: "#/definitions/valueObject.ContainerSpecs"
      name:
        type: string
      scalingIntervalSecs:
        type: integer
      scalingMaxDurationSecs:
        type: integer
      scalingPolicy:
        $ref: "#/definitions/valueObject.ScalingPolicy"
      scalingThreshold:
        type: integer
    type: object
...
  entity.ContainerProfile:
    properties:
      baseSpecs:
        $ref: "#/definitions/valueObject.ContainerSpecs"
      hostMinCapacityPercent:
        type: number
      id:
        type: integer
      maxSpecs:
        $ref: "#/definitions/valueObject.ContainerSpecs"
      name:
        type: string
      scalingIntervalSecs:
        type: integer
      scalingMaxDurationSecs:
        type: integer
      scalingPolicy:
        $ref: "#/definitions/valueObject.ScalingPolicy"
      scalingThreshold:
        type: integer
    type: object
  entity.O11yOverview:
    properties:
      hostname:
        type: string
      publicIp:
        type: string
      resourceUsage:
        $ref: "#/definitions/valueObject.HostResourceUsage"
      specs:
        $ref: "#/definitions/valueObject.HardwareSpecs"
      uptimeSecs:
        type: integer
    type: object
...
  valueObject.ContainerSpecs:
    properties:
      cpuCores:
        type: number
      memoryBytes:
        type: integer
    type: object
  valueObject.HardwareSpecs:
    properties:
      cpuCoresCount:
        type: number
      cpuFrequency:
        type: number
      cpuModelName:
        type: string
      memoryTotalBytes:
        type: integer
    type: object
  valueObject.HostResourceUsage:
    properties:
      cpuPercent:
        type: number
      memoryPercent:
        type: number
      netInfo:
        items:
          $ref: "#/definitions/valueObject.NetInterfaceInfo"
        type: array
      storageInfo:
        items:
          $ref: "#/definitions/valueObject.StorageUnitInfo"
        type: array
    type: object
  valueObject.NetInterfaceInfo:
    properties:
      deviceName:
        type: string
      recvBytes:
        type: integer
      recvDropPackets:
        type: integer
      recvErrs:
        type: integer
      recvPackets:
        type: integer
      sentBytes:
        type: integer
      sentDropPackets:
        type: integer
      sentErrs:
        type: integer
      sentPackets:
        type: integer
    type: object
...
  valueObject.ScalingPolicy:
    enum:
      - connection
      - cpu
      - memory
    type: string
    x-enum-varnames:
      - connection
      - cpu
      - memory
  valueObject.StorageUnitInfo:
    properties:
      deviceName:
        type: string
      fileSystem:
        type: string
      freeBytes:
        type: integer
      freeInodes:
        type: integer
      mountPoint:
        type: string
      readBytes:
        type: integer
      readOpsCount:
        type: integer
      totalBytes:
        type: integer
      totalInodes:
        type: integer
      usedBytes:
        type: integer
      usedInodes:
        type: integer
      usedInodesPercent:
        type: number
      usedPercent:
        type: number
      writeBytes:
        type: integer
      writeOpsCount:
        type: integer
    type: object
host: localhost:3141
info:
  contact:
    email: eng+swagger@speedia.net
    name: Infinite Engineering
    url: https://speedia.net/
  description: Infinite ControlManager API
  license:
    name: INFINITE CLOUD COMPUTING © 2023. All Rights Reserved.
    url: https://speedia.net/tos/
  termsOfService: https://speedia.net/tos/
  title: ControlApi
  version: 0.0.1
paths:
...
  /container/profile/:
    get:
      consumes:
        - application/json
      description: List container profiles.
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: "#/definitions/entity.ContainerProfile"
            type: array
      security:
        - Bearer: []
      summary: GetContainerProfiles
      tags:
        - container
    post:
      consumes:
        - application/json
      description: Add a new container profile.
      parameters:
        - description: NewContainerProfile (Only name and baseSpecs are required.)
          in: body
          name: addContainerProfileDto
          required: true
          schema:
            $ref: "#/definitions/dto.AddContainerProfile"
      produces:
        - application/json
      responses:
        "201":
          description: ContainerProfileCreated
          schema:
            type: object
      security:
        - Bearer: []
      summary: AddNewContainerProfile
      tags:
        - container
    put:
      consumes:
        - application/json
      description: Update a container profile.
      parameters:
        - description: UpdateContainerProfile (Only id is required.)
          in: body
          name: updateContainerProfileDto
          required: true
          schema:
            $ref: "#/definitions/dto.UpdateContainerProfile"
      produces:
        - application/json
      responses:
        "200":
          description: ContainerProfileUpdated
          schema:
            type: object
      security:
        - Bearer: []
      summary: UpdateContainerProfile
      tags:
        - container
  /container/profile/{profileId}/:
    delete:
      consumes:
        - application/json
      description: Delete a container profile.
      parameters:
        - description: ProfileId
          in: path
          name: profileId
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: ContainerProfileDeleted
          schema:
            type: object
      security:
        - Bearer: []
      summary: DeleteContainerProfile
      tags:
        - container
...
  /o11y/overview/:
    get:
      consumes:
        - application/json
      description: Show system information and resource usage.
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/entity.O11yOverview"
      security:
        - Bearer: []
      summary: O11yOverview
      tags:
        - o11y
securityDefinitions:
  Bearer:
    description: Type "Bearer" + JWT token or API key.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"